isolated sessionManagment
